Basic native counter app
